Bump up GHC-bundled dependencies to GHC@8.0.1
I've bumped them based on how the packages appear to be managed before,
so that the minimum versions are not less than the version included in
the minimum-supported GHC compiler (8.0.1).

Based on the table here: https://gitlab.haskell.org/ghc/ghc/wikis/commentary/libraries/version-history

Technically I think this might be unnecessary for several reasons:
  1. Since these packages are included in the minimum GHC release, I
  don't think earlier versions are even candidates for resolution.
  2. 8.0.1 is not formally supported, so perhaps these ought to be
  bumped all the way to the 8.0.2 versions.

Because of the first point, I think it is valid (and less error-prone)
to _remove_ the lower bound of these particular depdencies except where
they differ from the version included in the minimum supported GHC
version. I'm not that now because it might be a controversial change.